backend/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
pydantic
